Move login page to components
